renamed plugins file and check output exception
